requirements.txt
flask==3.1.0
python-dotenv==1.0.0
anthropic==0.7.0


PyPDF2==3.0.1
python-docx==1.1.0

